pyproject.toml
[project]
name = "jobtracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Allain Melisande",email = "melisande.allain@student-cs.fr"}
]
package-mode = false

requires-python = ">=3.9,<4.0"
dependencies = [
    "flask>=3.1.2,<4.0.0",
    "flask-cors>=4.0.0",
    "flask-sqlalchemy>=3.1.0",
    "pydantic>=2.12.5,<3.0.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "uuid (>=1.30,<2.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
